CancelOrder Service Added

+++ SmartMarket.Service/DTOs/CancelOrders/CancelOrderForResultDto.cs
@@ -10,7 +10,7 @@ public record CancelOrderForResultDto
     public string BarCode { get; set; }
     public long CategoryId { get; set; }
     public decimal Price { get; set; }
-    public short DiscountPrice { get; set; }
+    public decimal DiscountPrice { get; set; }
     public OlchovBirligi OlchovTuri { get; set; }
     public decimal Quantity { get; set; }
     public decimal TotalPrice { get; set; }
